refactor(backend): split securityValidators.ts (494) into 3 validators
Extract each middleware into its own single-responsibility file:

- validators/validateContactSubmission.ts (178 lines): contact form
  validation, XSS/SQL injection screening, field sanitization
- validators/validateSmartAlert.ts (193 lines): smart alert creation
  validation, criteria range checks, frequency enforcement
- validators/validateSearchParams.ts (122 lines): search param
  sanitization, geo-point validation, MongoDB operator stripping

securityValidators.ts reduced to 14-line re-export barrel.
Zero changes at the single import site.

File size: 494 -> 14 lines (97% reduction at entry point)
